% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.netglm.R
\name{gen.netglm}
\alias{gen.netglm}
\title{Simulate (generalized) linear network model data}
\usage{
gen.netglm(
  n,
  m,
  family = "gaussian",
  intercept = FALSE,
  beta = NULL,
  red.var = NULL,
  seed = NULL,
  ...
)
}
\arguments{
\item{n}{The size of the vertex set (|V(G)|) for the random graphs.}

\item{m}{The number of graphs to generate.}

\item{family}{Family of the generalized linear model used to generate the 
data. The available families are "gaussian" and "binomial".}

\item{intercept}{Logical; should the intercept be simulated? 
\code{FALSE} by default.}

\item{beta}{Values of the true coefficients used to generate the data. If 
\code{intercept} is \code{TRUE}, the first value in \code{beta} is considered
to be the intercept. If not provided, the coefficients are randomly sampled 
from a standard normal distribution and the intercept is defined as the maximum 
between \eqn{m/2} and the median of the beta coefficients. \code{NULL} by default.}

\item{red.var}{Value of the residual variance used when the family is "gaussian". 
If not provided, the residual variance is set to 1.}

\item{seed}{Integer used as the seed in the generation process. Allows the 
user to generate the same data consistently.}

\item{...}{other arguments passed on to the function \code{rgraph} of the 
\code{sna} package.}
}
\description{
Generates data based on a (generalized) linear network model, where the
dependent variable can be either continuous or dichotomous. All the 
independent variables are dichotomous.
}
