% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QAP.ML.R
\name{QAP.ML}
\alias{QAP.ML}
\title{Multilevel MRQAP}
\usage{
QAP.ML(
  formula,
  data,
  iv.list.per = "group",
  family = "gaussian",
  iv.names = iv.names,
  mode = "yQAP",
  samples = 1000,
  diag = F,
  directed = T,
  cpu = 1,
  round.to = 5,
  logfilename = "QAP.log",
  verbose = T,
  global.deltas = T,
  return.perms = F
)
}
\description{
Estimates a MRQAP model taking the multilevel/grouped nature of the into account.
}
\details{
#' @param family family of the generalized linear model. default is "gaussian" for continuous dependent variables. F
or binary dependent variables "binomial" is advised. (currently only gaussian or binomial possible)
}
