% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{QAP}
\alias{QAP}
\title{MRQAP for unnested data}
\usage{
QAP(dv, iv1, iv.names, mode = "yQAP", samples = 1000, diag = F, directed = F)
}
\arguments{
\item{dv}{a matrix with n * m dimensions and cells indicating the presence of a tie (1 = tie, 0 = no tie for binary variables) 
or the weight of a tie (for continous tie variables) charcterizing the dependent variable}

\item{iv1}{a list of matrices with n * m dimensions characterizing the independent variables}

\item{iv.names}{names of the independent variables for the output object}

\item{mode}{permutation method to be applied. default is "yQAP" for permuting the Y / dv
variables."dspQAP" applies Dekker's semi partialing method (Dekker, Krackhard, & Snijders, 2007)}

\item{samples}{number of permutations, default is 1000.}

\item{diag}{boolean for using the diagonal values of matrices in the estimation. default is FALSE}

\item{directed}{"directed" if the dependent network is directed (ties from A to B and B to A are possible), 
"undirected" if the dependent network is undirected (ties from A to B are identical to B to A). Default is "directed".}

\item{iv.list.per}{lists in the ivs argument are should be nested by group and independent matrices, if this is not the case (grouped by independent matrices and then groups)
the argument iv.list.per = "iv" can be used to restructure the data.}

\item{family}{family of the generalized linear model. default is "gaussian" for continuous dependent varaibles. F
or binday dependent variables "binomial" is advised.}

\item{round.to}{numeric, numer of digits in output table}

\item{cpu}{number of cpu's to be used for estimation, default is 1}

\item{logfilename}{name of log file printing intermediate reports during the estimation procedure.}

\item{verbose}{reports of what is happening under the hood during the call of the function, default is TRUE}

\item{global.deltas}{during "dspQAP" estimation, should global or local delta values be used. default is TRUE}

\item{return.perms}{should permuted networks be part of the output? default is FALSE}
}
\description{
Estimates a Multiple Regression Quadratic Assignment Proccedure model 
(MRQAP; Krackhardt, 1988). MRQAPs allow investigating associations between
 characteristics of dyads in networks (e.g., the level of homophily 
 between two actors) and a binary or continuous
  tie variable (e.g., friendship, amount of time spent together).
}
\examples{
# create test data #
inspired by the example funciton in sna::netlm
ivnet1<-sna::rgraph(20,4)

dv1<-ivnet1[1,,]+4*ivnet1[2,,]+2*ivnet1[3,,]   # Note that the fourth graph is unrelated
dv1 <- dv1 + rnorm(400,mean = 1, sd = 1)
iv1 <- list(ivnet1[1,,],ivnet1[2,,],ivnet1[3,,], ivnet1[4,,])
 QAP.MG(list(dv1), list(iv1), iv.names = c("intercept",paste0("IV",1:4)), samples = 3000)
}
\references{
Dekker, D.; Krackhardt, D.; Snijders, T.A.B.  (2007).  \dQuote{Sensitivity of MRQAP Tests to Collinearity and Autocorrelation Conditions.}  \emph{Psychometrika}, 72(4), 563-581.

Krackhardt, D.  (1987).  \dQuote{QAP Partialling as a Test of Spuriousness.} \emph{Social Networks}, 9 171-186.

Krackhardt, D.  (1988).  \dQuote{Predicting With Networks: Nonparametric Multiple Regression Analyses of Dyadic Data.}  \emph{Social Networks}, 10, 359-382.
}
\seealso{
\code{\link{QAP.MG}}
}
